pyproject.toml
[project]
name = "mcp-python-demo"
version = "0.1.0"
description = "MCP Python Demo with Web UI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "openai>=1.73.0",
    "streamlit>=1.24.0",
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "nest-asyncio>=1.5.6",
    "websockets>=10.0",
    "python-dotenv>=0.19.0",
    "asyncclick>=8.0.1",
    "zhipuai>=1.0.7",
    "anyio>=4.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true